backend/requirements.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
bqplot==0.12.45
branca==0.8.1
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==1.0.15
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpathlib==0.21.1
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
crewai==0.140.0
crewai-tools==0.49.0
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.10
earthengine-api==1.5.22
eerepr==0.1.2
embedchain==0.1.128
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
fastavro==1.11.1
filelock==3.18.0
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
geemap==0.35.3
geocoder==1.38.1
geographiclib==2.0
geopandas==1.1.1
geopy==2.4.1
ghp-import==2.1.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.3
groq==0.29.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.2
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.9.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.20.0
ipython==8.37.0
ipytree==0.2.2
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-leaflet==0.20.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==33.1.0
lancedb==0.24.0
langchain==0.3.26
langchain-cohere==0.3.5
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-groq==0.3.5
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.3.45
language_data==1.3.0
litellm==1.72.6
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.114
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.1
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
portalocker==2.10.1
posthog==5.4.0
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16
pyogrio==0.11.0
pyparsing==3.2.3
pypdf==5.7.0
pypdfium2==4.30.1
pyperclip==1.9.0
PyPika==0.48.9
pyproj==3.7.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.402
pysbd==0.3.4
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
pywin32==310
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==27.0.0
qdrant-client==1.14.3
rasterio==1.4.3
ratelim==0.1.6
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
schema==0.7.7
scooby==0.10.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.19
uvicorn==0.35.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
